Finished Storkey Rule implementation and basis unit tests to confirm the weight matrix is being calculated correctly and certain simple exemplars are returned correctly.
Added the "scaled" flag for indicating if the Hebbian Learning Rule should be executed as defined in Storkeys paper (each weight updated scaled by 1/n) or as we learned in the slides for module 8 (no 1/n factor).

Updated Jupyter Notebook to execute the Storkey Learning Rule on the exemplars, followed with synchronous and asynchronous recovery results.